* Fixed warnings in new CertificateFinder* and Ldap/Downloader code
git-svn-id: svn://svn.minisip.org/minisip/trunk@3365 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmcrypto/source/CertificatePathFinderUcd.cxx
@@ -251,7 +251,7 @@ std::vector<MRef<Certificate*> > CertificatePathFinderUcd::findCerts	(std::vecto
 		which of the candidates that should be tested first) exceed the number of
 		candidates then we should abort the function because there is nothing more to be done.
 		*/
-		if (candidates.empty() || phaseEffort >= candidates.size()) {
+		if (candidates.empty() || phaseEffort >= (int)candidates.size()) {
 			phaseEffort = MAX_EFFORT;
 			findEffort = MAX_EFFORT;
 			break;
